+++ internal/provider/resource_sdwan_transport_wan_vpn_interface_t1_e1_serial_profile_parcel.go
@@ -1,0 +1,746 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+	"net/url"
+	"regexp"
+	"sync"
+
+	"github.com/CiscoDevNet/terraform-provider-sdwan/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/netascode/go-sdwan"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin model
+
+// Ensure provider defined types fully satisfy framework interfaces
+var _ resource.Resource = &TransportWANVPNInterfaceT1E1SerialProfileParcelResource{}
+var _ resource.ResourceWithImportState = &TransportWANVPNInterfaceT1E1SerialProfileParcelResource{}
+
+func NewTransportWANVPNInterfaceT1E1SerialProfileParcelResource() resource.Resource {
+	return &TransportWANVPNInterfaceT1E1SerialProfileParcelResource{}
+}
+
+type TransportWANVPNInterfaceT1E1SerialProfileParcelResource struct {
+	client      *sdwan.Client
+	updateMutex *sync.Mutex
+}
+
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_transport_wan_vpn_interface_t1_e1_serial_profile_parcel"
+}
+
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: helpers.NewAttributeDescription("This resource can manage a Transport WAN VPN Interface T1 E1 Serial profile parcel.").AddMinimumVersionDescription("20.12.0").String,
+
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{
+				MarkdownDescription: "The id of the profile parcel",
+				Computed:            true,
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.UseStateForUnknown(),
+				},
+			},
+			"version": schema.Int64Attribute{
+				MarkdownDescription: "The version of the profile parcel",
+				Computed:            true,
+			},
+			"name": schema.StringAttribute{
+				MarkdownDescription: "The name of the profile parcel",
+				Required:            true,
+			},
+			"description": schema.StringAttribute{
+				MarkdownDescription: "The description of the profile parcel",
+				Optional:            true,
+			},
+			"feature_profile_id": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Feature Profile ID").String,
+				Required:            true,
+			},
+			"transport_wan_vpn_profile_parcel_id": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Transport WAN VPN Profile Parcel ID").String,
+				Optional:            true,
+			},
+			"shutdown": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Administrative state").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"shutdown_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"interface_name": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Serial Interface Name - slot/subslot/port:channel-group for T1/E1, slot/subslot/port for NIM-1T").String,
+				Required:            true,
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(7, 255),
+					stringvalidator.RegexMatches(regexp.MustCompile(`^(Serial)([0-9]{1,}(. ?[1-9][0-9]*)*|[0-9/]+|[0-9]+/[0-9]+/[0-9]+:[0-9]+|[0-9]+/[0-9]+/[0-9]+|[0-9]+/[0-9]+|[0-9]+)`), ""),
+				},
+			},
+			"interface_name_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"ipv4_address": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("").String,
+				Optional:            true,
+			},
+			"ipv4_address_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"ipv4_mask": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("").AddStringEnumDescription("255.255.255.255", "255.255.255.254", "255.255.255.252", "255.255.255.248", "255.255.255.240", "255.255.255.224", "255.255.255.192", "255.255.255.128", "255.255.255.0", "255.255.254.0", "255.255.252.0", "255.255.248.0", "255.255.240.0", "255.255.224.0", "255.255.192.0", "255.255.128.0", "255.255.0.0", "255.254.0.0", "255.252.0.0", "255.240.0.0", "255.224.0.0", "255.192.0.0", "255.128.0.0", "255.0.0.0", "254.0.0.0", "252.0.0.0", "248.0.0.0", "240.0.0.0", "224.0.0.0", "192.0.0.0", "128.0.0.0", "0.0.0.0").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("255.255.255.255", "255.255.255.254", "255.255.255.252", "255.255.255.248", "255.255.255.240", "255.255.255.224", "255.255.255.192", "255.255.255.128", "255.255.255.0", "255.255.254.0", "255.255.252.0", "255.255.248.0", "255.255.240.0", "255.255.224.0", "255.255.192.0", "255.255.128.0", "255.255.0.0", "255.254.0.0", "255.252.0.0", "255.240.0.0", "255.224.0.0", "255.192.0.0", "255.128.0.0", "255.0.0.0", "254.0.0.0", "252.0.0.0", "248.0.0.0", "240.0.0.0", "224.0.0.0", "192.0.0.0", "128.0.0.0", "0.0.0.0"),
+				},
+			},
+			"ipv4_mask_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"ipv6_address": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Assign IPv6 address").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.RegexMatches(regexp.MustCompile(`((^\s*((([0-9A-Fa-f]{1,4}:){7}([0-9A-Fa-f]{1,4}|:))|(([0-9A-Fa-f]{1,4}:){6}(:[0-9A-Fa-f]{1,4}|((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3})|:))|(([0-9A-Fa-f]{1,4}:){5}(((:[0-9A-Fa-f]{1,4}){1,2})|:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3})|:))|(([0-9A-Fa-f]{1,4}:){4}(((:[0-9A-Fa-f]{1,4}){1,3})|((:[0-9A-Fa-f]{1,4})?:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){3}(((:[0-9A-Fa-f]{1,4}){1,4})|((:[0-9A-Fa-f]{1,4}){0,2}:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){2}(((:[0-9A-Fa-f]{1,4}){1,5})|((:[0-9A-Fa-f]{1,4}){0,3}:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){1}(((:[0-9A-Fa-f]{1,4}){1,6})|((:[0-9A-Fa-f]{1,4}){0,4}:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3}))|:))|(:(((:[0-9A-Fa-f]{1,4}){1,7})|((:[0-9A-Fa-f]{1,4}){0,5}:((25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)(\.(25[0-5]|2[0-4]\d|1\d\d|[1-9]?\d)){3}))|:)))(%.+)?\s*(/)(\b([0-9]{1,2}|1[01][0-9]|12[0-8])\b)$))`), ""),
+				},
+			},
+			"ipv6_address_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"bandwidth": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Interface bandwidth capacity, in kbps").AddIntegerRangeDescription(1, 200000000).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(1, 200000000),
+				},
+			},
+			"bandwidth_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"bandwidth_downstream": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Interface downstream bandwidth capacity, in kbps").AddIntegerRangeDescription(1, 2147483647).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(1, 2147483647),
+				},
+			},
+			"bandwidth_downstream_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"clock_rate": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set preference for interface Clock speed").AddStringEnumDescription("1200", "2400", "4800", "9600", "14400", "19200", "28800", "32000", "38400", "48000", "56000", "57600", "64000", "72000", "115200", "125000", "148000", "192000", "250000", "256000", "384000", "500000", "512000", "768000", "800000", "1000000", "2000000", "4000000", "5300000", "8000000").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("1200", "2400", "4800", "9600", "14400", "19200", "28800", "32000", "38400", "48000", "56000", "57600", "64000", "72000", "115200", "125000", "148000", "192000", "250000", "256000", "384000", "500000", "512000", "768000", "800000", "1000000", "2000000", "4000000", "5300000", "8000000"),
+				},
+			},
+			"clock_rate_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"encapsulation": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Configure Encapsulation for interface").AddStringEnumDescription("hdlc", "ppp", "frame-relay").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("hdlc", "ppp", "frame-relay"),
+				},
+			},
+			"encapsulation_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Tunnel Interface").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"per_tunnel_qos": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Per-tunnel Qos").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"per_tunnel_qos_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"per_tunnel_qos_aggregator": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Per-tunnel QoS Aggregator").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"per_tunnel_qos_aggregator_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_qos_mode": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set tunnel QoS mode").AddStringEnumDescription("spoke", "hub").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("spoke", "hub"),
+				},
+			},
+			"tunnel_qos_mode_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_color": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set color for TLOC").AddStringEnumDescription("default", "mpls", "metro-ethernet", "biz-internet", "public-internet", "lte", "3g", "red", "green", "blue", "gold", "silver", "bronze", "custom1", "custom2", "custom3", "private1", "private2", "private3", "private4", "private5", "private6").AddDefaultValueDescription("default").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("default", "mpls", "metro-ethernet", "biz-internet", "public-internet", "lte", "3g", "red", "green", "blue", "gold", "silver", "bronze", "custom1", "custom2", "custom3", "private1", "private2", "private3", "private4", "private5", "private6"),
+				},
+			},
+			"tunnel_interface_color_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_color_restrict": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Restrict this TLOC behavior").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_color_restrict_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_groups": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("List of groups").AddIntegerRangeDescription(1, 4294967295).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(1, 4294967295),
+				},
+			},
+			"tunnel_interface_groups_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_border": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set TLOC as border TLOC").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_border_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_max_control_connections": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set the maximum number of control connections for this TLOC").AddIntegerRangeDescription(0, 100).String,
+				Optional:            true,
+			},
+			"tunnel_interface_max_control_connections_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_vbond_as_stun_server": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Put this wan interface in STUN mode only").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_vbond_as_stun_server_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_exclude_controller_group_list": schema.SetAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Exclude the following controller groups defined in this list").String,
+				ElementType:         types.Int64Type,
+				Optional:            true,
+			},
+			"tunnel_interface_exclude_controller_group_list_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_vmanage_connection_preference": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set interface preference for control connection to vManage <0..8>").AddIntegerRangeDescription(0, 8).AddDefaultValueDescription("5").String,
+				Optional:            true,
+			},
+			"tunnel_interface_vmanage_connection_preference_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_port_hop": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Disallow port hopping on the tunnel interface").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"tunnel_interface_port_hop_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_low_bandwidth_link": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set the interface as a low-bandwidth circuit").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_low_bandwidth_link_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_tunnel_tcp_mss": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Tunnel TCP MSS on SYN packets, in bytes").AddIntegerRangeDescription(500, 1460).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(500, 1460),
+				},
+			},
+			"tunnel_interface_tunnel_tcp_mss_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_clear_dont_fragment": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Enable clear dont fragment (Currently Only SDWAN Tunnel Interface)").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_clear_dont_fragment_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_clear_network_broadcast": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Accept and respond to network-prefix-directed broadcasts)").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_clear_network_broadcast_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_carrier": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set carrier for TLOC").AddStringEnumDescription("default", "carrier1", "carrier2", "carrier3", "carrier4", "carrier5", "carrier6", "carrier7", "carrier8").AddDefaultValueDescription("default").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.OneOf("default", "carrier1", "carrier2", "carrier3", "carrier4", "carrier5", "carrier6", "carrier7", "carrier8"),
+				},
+			},
+			"tunnel_interface_carrier_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_bind_loopback_tunnel": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Bind loopback tunnel interface to a physical interface").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(1, 32),
+				},
+			},
+			"tunnel_interface_bind_loopback_tunnel_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_last_resort_circuit": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set TLOC as last resort").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_last_resort_circuit_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_nat_refresh_interval": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set time period of nat refresh packets <1...60> seconds").AddIntegerRangeDescription(1, 60).AddDefaultValueDescription("5").String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(1, 60),
+				},
+			},
+			"tunnel_interface_nat_refresh_interval_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_hello_interval": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set time period of control hello packets <100..600000> milli seconds").AddIntegerRangeDescription(100, 600000).AddDefaultValueDescription("1000").String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(100, 600000),
+				},
+			},
+			"tunnel_interface_hello_interval_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_hello_tolerance": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set tolerance of control hello packets <12..6000> seconds").AddIntegerRangeDescription(12, 6000).AddDefaultValueDescription("12").String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(12, 6000),
+				},
+			},
+			"tunnel_interface_hello_tolerance_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_all": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow all traffic. Overrides all other allow-service options if allow-service all is set").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_all_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_bgp": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/deny BGP").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_bgp_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_dhcp": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny DHCP").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_dhcp_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_dbs": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny DNS").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_dbs_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_icmp": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny ICMP").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_icmp_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_netconf": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny NETCONF").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_netconf_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ntp": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny NTP").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ntp_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ospf": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny OSPF").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ospf_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ssh": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny SSH").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_ssh_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_stun": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny STUN").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_stun_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_https": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny Https").AddDefaultValueDescription("true").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_https_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_snmp": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny SNMP").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_snmp_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_bfd": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Allow/Deny BFD").AddDefaultValueDescription("false").String,
+				Optional:            true,
+			},
+			"tunnel_interface_allow_bfd_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tunnel_interface_encapsulations": schema.ListNestedAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Encapsulation for TLOC").String,
+				Optional:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"encapsulation": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Encapsulation").AddStringEnumDescription("gre", "ipsec").String,
+							Optional:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("gre", "ipsec"),
+							},
+						},
+						"preference": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Set preference for TLOC").AddIntegerRangeDescription(0, 4294967295).String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.AtMost(4294967295),
+							},
+						},
+						"preference_variable": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+							Optional:            true,
+						},
+						"weight": schema.Int64Attribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Set weight for TLOC").AddIntegerRangeDescription(1, 255).AddDefaultValueDescription("1").String,
+							Optional:            true,
+							Validators: []validator.Int64{
+								int64validator.Between(1, 255),
+							},
+						},
+						"weight_variable": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+							Optional:            true,
+						},
+					},
+				},
+			},
+			"acl_shaping_rate": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("1ge  interfaces: [0..1000000]kbps; 10ge interfaces: [0..10000000]kbps").AddIntegerRangeDescription(8, 100000000).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(8, 100000000),
+				},
+			},
+			"acl_shaping_rate_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tcp_mss": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("TCP MSS on SYN packets, in bytes").AddIntegerRangeDescription(500, 1460).String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(500, 1460),
+				},
+			},
+			"tcp_mss_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"mtu": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Interface MTU <68...2000>, in bytes").AddIntegerRangeDescription(576, 9216).AddDefaultValueDescription("1500").String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(576, 9216),
+				},
+			},
+			"mtu_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"ip_mtu": schema.Int64Attribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Set ip mtu").AddIntegerRangeDescription(576, 9216).AddDefaultValueDescription("1500").String,
+				Optional:            true,
+				Validators: []validator.Int64{
+					int64validator.Between(576, 9216),
+				},
+			},
+			"ip_mtu_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+			"tloc_extension": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Extends a local TLOC to a remote node only for vpn 0").String,
+				Optional:            true,
+				Validators: []validator.String{
+					stringvalidator.LengthBetween(1, 32),
+				},
+			},
+			"tloc_extension_variable": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Variable name").String,
+				Optional:            true,
+			},
+		},
+	}
+}
+
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Configure(_ context.Context, req resource.ConfigureRequest, _ *resource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	r.client = req.ProviderData.(*SdwanProviderData).Client
+	r.updateMutex = req.ProviderData.(*SdwanProviderData).UpdateMutex
+}
+
+// End of section. //template:end model
+
+// Section below is generated&owned by "gen/generator.go". //template:begin create
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var plan TransportWANVPNInterfaceT1E1Serial
+
+	// Read plan
+	diags := req.Plan.Get(ctx, &plan)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Create", plan.Name.ValueString()))
+
+	// Create object
+	body := plan.toBody(ctx)
+
+	res, err := r.client.Post(plan.getPath(), body)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST), got error: %s, %s", err, res.String()))
+		return
+	}
+
+	plan.Id = types.StringValue(res.Get("parcelId").String())
+	plan.Version = types.Int64Value(0)
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished successfully", plan.Name.ValueString()))
+
+	diags = resp.State.Set(ctx, &plan)
+	resp.Diagnostics.Append(diags...)
+}
+
+// End of section. //template:end create
+
+// Section below is generated&owned by "gen/generator.go". //template:begin read
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var state TransportWANVPNInterfaceT1E1Serial
+
+	// Read state
+	diags := req.State.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", state.Name.String()))
+
+	res, err := r.client.Get(state.getPath() + "/" + url.QueryEscape(state.Id.ValueString()))
+	if res.Get("error.message").String() == "Invalid feature Id" {
+		resp.State.RemoveResource(ctx)
+		return
+	} else if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, res.String()))
+		return
+	}
+
+	// If every attribute is set to null we are dealing with an import operation and therefore reading all attributes
+	if state.isNull(ctx, res) {
+		state.fromBody(ctx, res)
+	} else {
+		state.updateFromBody(ctx, res)
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Name.ValueString()))
+
+	diags = resp.State.Set(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+}
+
+// End of section. //template:end read
+
+// Section below is generated&owned by "gen/generator.go". //template:begin update
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan, state TransportWANVPNInterfaceT1E1Serial
+
+	// Read plan
+	diags := req.Plan.Get(ctx, &plan)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+	// Read state
+	diags = req.State.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Update", plan.Name.ValueString()))
+
+	body := plan.toBody(ctx)
+	res, err := r.client.Put(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString()), body)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (PUT), got error: %s, %s", err, res.String()))
+		return
+	}
+
+	plan.Version = types.Int64Value(state.Version.ValueInt64() + 1)
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Update finished successfully", plan.Name.ValueString()))
+
+	diags = resp.State.Set(ctx, &plan)
+	resp.Diagnostics.Append(diags...)
+}
+
+// End of section. //template:end update
+
+// Section below is generated&owned by "gen/generator.go". //template:begin delete
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var state TransportWANVPNInterfaceT1E1Serial
+
+	// Read state
+	diags := req.State.Get(ctx, &state)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Delete", state.Name.ValueString()))
+
+	res, err := r.client.Delete(state.getPath() + "/" + url.QueryEscape(state.Id.ValueString()))
+	if err != nil && res.Get("error.message").String() != "Invalid Template Id" {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object (DELETE), got error: %s, %s", err, res.String()))
+		return
+	}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Delete finished successfully", state.Name.ValueString()))
+
+	resp.State.RemoveResource(ctx)
+}
+
+// End of section. //template:end delete
+
+// Section below is generated&owned by "gen/generator.go". //template:begin import
+func (r *TransportWANVPNInterfaceT1E1SerialProfileParcelResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
+	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
+}
+
+// End of section. //template:end import