+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `sdwan_transport_wan_vpn_interface_ethernet_profile_parcel` resource and data source
 - Add `sdwan_transport_wan_vpn_interface_gre_profile_parcel` resource and data source
 - Add `sdwan_transport_wan_vpn_interface_ipsec_profile_parcel` resource and data source
+- Add `sdwan_transport_wan_vpn_interface_t1_e1_serial_profile_parcel` resource and data source
 
 ## 0.3.9